infosite/requirements.txt
Pillow==8.1.0
Django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-ckeditor==6.0.0

